Laser inline power working.
There remains a bug where if the laser fan is not on the first CAN send will be triggered from the block_phase ISR.

CanChannel:Write will take a semaphore.
The semaphore will call taskENTER_CRITICAL().
taskENTER_CRITICAL cannot be called from within an ISR and will fail a Free_RTOS assertion, causing a hang.

+++ Marlin/src/module/planner.cpp
@@ -850,7 +850,7 @@ void Planner::calculate_trapezoid_for_block(block_t* const block, const float &e
    */
   #if ENABLED(LASER_POWER_INLINE_TRAPEZOID)
     if (block->laser.power > 0) { // No need to care if power == 0
-      const uint8_t entry_power = block->laser.power * entry_factor; // Power on block entry
+      const float entry_power = block->laser.power * entry_factor; // Power on block entry
       #if DISABLED(LASER_POWER_INLINE_TRAPEZOID_CONT)
         // Speedup power
         const uint8_t entry_power_diff = block->laser.power - entry_power;
